rename css file more concisely
